# cmd/api/docs/swagger.yaml
definitions:
...
  requests.OTPreq:
    properties:
      Phone:
        type: string
    required:
    - Phone
    type: object
  requests.Otpverifier:
    properties:
      Phone:
        type: string
      pin:
        type: string
    required:
    - Phone
    - pin
    type: object
...
  requests.Update:
    properties:
      order_id:
        type: integer
      status_id:
        type: integer
    required:
    - order_id
    - status_id
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      stastuscode:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /admin/order/Allorders:
    get:
      consumes:
      - application/json
      operationId: FindAllordersInshop
      parameters:
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Number of items to retrieve per page
        in: query
        name: perPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: FindAllorders In admin side
      tags:
      - Order
  /admin/order/Status:
    get:
      consumes:
      - application/json
      description: Endpoint for getting all orderStatuses
      operationId: List-all-orderStatus
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: for geting all order status list
      tags:
      - Order
  /admin/order/UpdateStatus:
    patch:
      consumes:
      - application/json
      description: update the order statuses by every orderid.
      operationId: Order_status
      parameters:
      - description: Input Field
        in: body
        name: inputs
        required: true
        schema:
          $ref: '#/definitions/requests.Update'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      summary: Updateorderstatus
      tags:
      - Order
...
  /order/cancel/{orderId}:
    patch:
      consumes:
      - application/json
      description: Endpoint for cancelling an order associated with a user
      operationId: cancel-order
      parameters:
      - description: ID of the order to be cancelled
        in: path
        name: orderId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Cancels a specific order for the currently logged in user
      tags:
      - Order
  /order/listall:
    get:
      consumes:
      - application/json
      description: Endpoint for getting all orders
      operationId: List-all-orders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: for geting all order list
      tags:
      - Order
  /order/orderAll/{payment_id}:
    post:
      consumes:
      - application/json
      description: This endpoint allows a user to purchase all items in their cart
      operationId: buyAll
      parameters:
      - description: payment_id
        in: path
        name: payment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Buy all items from the user's cart
      tags:
      - Order
  /order/return/{orderId}:
    patch:
      consumes:
      - application/json
      description: Endpoint for Returning an order associated with a user
      operationId: return-order
      parameters:
      - description: ID of the order to be cancelled
        in: path
        name: orderId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Return a specific order for the currently logged in user
      tags:
      - Order
  /order/view/{order_id}:
    get:
      consumes:
      - application/json
      description: retrieving the details of a specific order identified by its order
        ID.
      operationId: view-order-by-id
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched order details
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Failed to fetch order details
          schema:
            $ref: '#/definitions/response.Response'
      summary: to get order details
      tags:
      - Order
  /otp/send:
    post:
      consumes:
      - application/json
      description: Send OTP to use's mobile
      operationId: send-otp
      parameters:
      - description: User mobile number
        in: body
        name: user_mobile
        required: true
        schema:
          $ref: '#/definitions/requests.OTPreq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Send OTP to user's mobile
      tags:
      - Otp
  /otp/verify:
    post:
      consumes:
      - application/json
      description: Validate the  OTP sent to use's mobile
      operationId: validate-otp
      parameters:
      - description: OTP sent to user's mobile number
        in: body
        name: otp
        required: true
        schema:
          $ref: '#/definitions/requests.Otpverifier'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Validate the OTP to user's mobile
      tags:
      - Otp
...
swagger: "2.0"